# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSA Backend - All Projects",
    "version": "1.0.0",
    "description": "Every endpoint of every project backend in this repository. Operations are tagged and prefixed with the project they belong to."
  },
  "paths": {
...
    "/membership-analytics/api/v1/cases/overview": {
      "get": {
        "tags": [
          "Membership Analytics - Cases Dashboard"
        ],
        "summary": "[Membership Analytics] Overview",
        "operationId": "Membership_Analytics__overview_api_v1_cases_overview_get",
        "parameters": [
          {
            "name": "parliament",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parliament"
            }
          },
          {
            "name": "constituency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Constituency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/cases/leaders": {
      "get": {
        "tags": [
          "Membership Analytics - Cases Dashboard"
        ],
        "summary": "[Membership Analytics] Leaders",
        "operationId": "Membership_Analytics__leaders_api_v1_cases_leaders_get",
        "parameters": [
          {
            "name": "parliament",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parliament"
            }
          },
          {
            "name": "constituency",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Constituency"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(leaders|all)$",
              "default": "leaders",
              "title": "Scope"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/cases/search": {
      "get": {
        "tags": [
          "Membership Analytics - Cases Dashboard"
        ],
        "summary": "[Membership Analytics] Search",
        "operationId": "Membership_Analytics__search_api_v1_cases_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/cases/leaders/{leader_key}": {
      "get": {
        "tags": [
          "Membership Analytics - Cases Dashboard"
        ],
        "summary": "[Membership Analytics] Leader Detail",
        "operationId": "Membership_Analytics__leader_detail_api_v1_cases_leaders__leader_key__get",
        "parameters": [
          {
            "name": "leader_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Leader Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Membership_Analytics__HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/Membership_Analytics__ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Membership_Analytics__ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Membership_Analytics__app__schemas__nominated_schema__ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "OK"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse"
      },
...
    }
  },
...
}